+++ samples/stellar-cartpole-trpo/cartpole.py
@@ -5,15 +5,14 @@ permalink: https://perma.cc/C9ZM-652R
 """
 
 import math
-import numpy as np
-import seeding
-from bonsai_ai.logger import Logger
+import random
+from collections import namedtuple
 
-log = Logger()
+CartPoleState = namedtuple('CartPoleState', 'x x_dot y y_dot')
 
-
-class CartPoleModel():
-
+class CartPole:
+    """ Model for the dynamics of an inverted pendulum
+    """
     def __init__(self):
         self.gravity = 9.8
         self.masscart = 1.0
@@ -24,28 +23,19 @@ class CartPoleModel():
         self.force_mag = 10.0
         self.tau = 0.02  # seconds between state updates
 
-        # Angle at which to fail the episode
         self.theta_threshold_radians = 12 * 2 * math.pi / 360
         self.x_threshold = 2.4
+        
+        self.reset()
 
-        self._seed()
-        self.viewer = None
-        self.state = None
-
-        self.steps_beyond_done = None
-
-    def _seed(self, seed=None):
-        self.np_random, seed = seeding.np_random(seed)
-        return [seed]
-
-    def _step(self, action):
-        state = self.state
-        x, x_dot, theta, theta_dot = state
-        force = self.force_mag if action == 1 else -self.force_mag
-        costheta = math.cos(theta)
-        sintheta = math.sin(theta)
+    def step(self, action):
+        """ Move the state of the cartpole simulation forward one time unit
+        """
+        force = self.force_mag if action else -self.force_mag
+        costheta = math.cos(self.theta)
+        sintheta = math.sin(self.theta)
         temp = (
-            (force + self.polemass_length * theta_dot * theta_dot * sintheta)
+            (force + self.polemass_length * self.theta_dot ** 2 * sintheta)
             / self.total_mass)
         thetaacc = (
             (self.gravity * sintheta - costheta * temp)
@@ -55,38 +45,22 @@ class CartPoleModel():
         xacc = (
             temp - self.polemass_length * thetaacc * costheta / self.total_mass
         )
-        x = x + self.tau * x_dot
-        x_dot = x_dot + self.tau * xacc
-        theta = theta + self.tau * theta_dot
-        theta_dot = theta_dot + self.tau * thetaacc
-        self.state = (x, x_dot, theta, theta_dot)
-        done = x < -self.x_threshold \
-            or x > self.x_threshold \
-            or theta < -self.theta_threshold_radians \
-            or theta > self.theta_threshold_radians
-        done = bool(done)
+        self.x += self.tau * self.x_dot
+        self.x_dot += self.tau * xacc
+        self.theta += self.tau * self.theta_dot
+        self.theta_dot += self.tau * thetaacc
 
-        if not done:
-            reward = 1.0
-        elif self.steps_beyond_done is None:
-            # Pole just fell!
-            self.steps_beyond_done = 0
-            reward = 1.0
-        else:
-            if self.steps_beyond_done == 0:
-                log.warning("You are calling 'step()' even though this \
-                environment has already returned done = True. You should \
-                always call 'reset()' once you receive 'done = True' -- \
-                any further steps are undefined behavior.")
-            self.steps_beyond_done += 1
-            reward = 0.0
+        return self.state
 
-        return (np.array(self.state),
-                reward,
-                done,
-                {"steps_beyond_done": self.steps_beyond_done})
+    def reset(self):
+        """ Reset the model of a cartpole system to it's initial conditions
+        """
+        self.x = random.uniform(-0.05, 0.05)
+        self.x_dot = random.uniform(-0.05, 0.05)
+        self.theta = random.uniform(-0.05, 0.05)
+        self.theta_dot = random.uniform(-0.05, 0.05)
+    
+    @property
+    def state(self):
+        return CartPoleState(self.x, self.x_dot, self.theta, self.theta_dot)
 
-    def _reset(self):
-        self.state = self.np_random.uniform(low=-0.05, high=0.05, size=(4,))
-        self.steps_beyond_done = None
-        return np.array(self.state)