+++ samples/stellar-cartpole-trpo/render.py
@@ -5,68 +5,123 @@ permalink: https://perma.cc/C9ZM-652R
 """
 
 import math
+import pyglet
+from pyglet.gl import (gl, glBegin, glBlendFunc, glClearColor, glColor4f, 
+                       glEnable, glEnd, glLineWidth, glPopMatrix, glPushMatrix, 
+                       glRotatef, glTranslatef, glVertex2f, glVertex3f, 
+                       GL_BLEND, GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA)
+
+import six
+
+RAD2DEG = 57.29577951308232
+
+def draw_cartpole(cartpole, screen_width):
+    """ Draws a cartpole on a screen, scaled to screen_width
+    """
+    # Compute size of cart and pole as a function of screen_width
+    world_width = cartpole.x_threshold*2
+    scale = screen_width/world_width
+    carty = 100  # TOP OF CART
+    polewidth = 10.0
+    polelen = scale * 1.0
+    cartwidth = 50.0
+    cartheight = 30.0
+    
+    # Draw track
+    glColor4f(0,0,0,1.0)
+    glLineWidth(1.0)
+    glBegin(gl.GL_LINES)
+    glVertex2f(0, carty)
+    glVertex2f(screen_width, carty)
+    glEnd()
+
+    # Draw Cart
+    l, r, t, b = -cartwidth/2, cartwidth/2, cartheight/2, -cartheight/2
+    cartx = cartpole.x*scale+screen_width/2.0  # MIDDLE OF CART
+    
+    glColor4f(0.,0.,0.,1.0)
+    glPushMatrix()  # Push Translation
+    glTranslatef(cartx, carty, 0)
+    glBegin(gl.GL_QUADS)
+    glVertex3f(l, b, 0)
+    glVertex3f(l, t, 0)
+    glVertex3f(r, t, 0)
+    glVertex3f(r, b, 0)
+    glEnd()
+    
+    # Draw Pole
+    l, r, t, b = (
+        -polewidth/2,
+        polewidth/2,
+        polelen-polewidth/2,
+        -polewidth/2)
+    
+    glColor4f(.8, .6, .4, 1.0)
+    glPushMatrix()  # Push Rotation
+    glRotatef(RAD2DEG * -cartpole.theta, 0, 0, 1.0)
+    glBegin(gl.GL_QUADS)
+    glVertex3f(l, b, 0)
+    glVertex3f(l, t, 0)
+    glVertex3f(r, t, 0)
+    glVertex3f(r, b, 0)
+    glEnd()
+    glPopMatrix()  # Pop Rotation
+    
+    # Draw Axle
+    radius = polewidth/2
+    glColor4f(0.5, 0.5, 0.8, 1.0)
+    glBegin(gl.GL_POLYGON)
+    for i in range(12):
+        ang = 2 * math.pi * i / 12
+        x=math.cos(ang)*radius
+        y=math.sin(ang)*radius
+        glVertex3f(x, y, 0)
+    glEnd()
+    glPopMatrix()  # Pop Translation
 
 
-class _renderer():
+class Viewer(pyglet.window.Window):
+    """ Window for rendering a visualization of the CartPole model.
+    """
+    def __init__(self, width=600, height=400, display=None):
+        display = self._get_display(display)
+        super().__init__(width=width, height=height, display=display)
+        
+        self.model = None
 
-    def __init__(self, model, mode='human', close=False):
-        self.viewer = None
-        self.model = model
+        glEnable(GL_BLEND)
+        glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA)
+    
+    def _get_display(self, spec):
+        """Convert a display specification (such as :0) into an actual Display
+        object.
 
-    def _render(self, mode='human', close=False):
-        if close:
-            if self.viewer is not None:
-                self.viewer.close()
-                self.viewer = None
-            return
-
-        screen_width = 600
-        screen_height = 400
-
-        world_width = self.model.x_threshold*2
-        scale = screen_width/world_width
-        carty = 100  # TOP OF CART
-        polewidth = 10.0
-        polelen = scale * 1.0
-        cartwidth = 50.0
-        cartheight = 30.0
-
-        if self.viewer is None:
-            import rendering
-            # rom gym.envs.classic_control import rendering
-            self.viewer = rendering.Viewer(screen_width, screen_height)
-            l, r, t, b = -cartwidth/2, cartwidth/2, cartheight/2, -cartheight/2
-            axleoffset = cartheight/4.0
-            cart = rendering.FilledPolygon([(l, b), (l, t), (r, t), (r, b)])
-            self.carttrans = rendering.Transform()
-            cart.add_attr(self.carttrans)
-            self.viewer.add_geom(cart)
-            l, r, t, b = (
-                -polewidth/2,
-                polewidth/2,
-                polelen-polewidth/2,
-                -polewidth/2)
-            pole = rendering.FilledPolygon([(l, b), (l, t), (r, t), (r, b)])
-            pole.set_color(.8, .6, .4)
-            self.poletrans = rendering.Transform(translation=(0, axleoffset))
-            pole.add_attr(self.poletrans)
-            pole.add_attr(self.carttrans)
-            self.viewer.add_geom(pole)
-            self.axle = rendering.make_circle(polewidth/2)
-            self.axle.add_attr(self.poletrans)
-            self.axle.add_attr(self.carttrans)
-            self.axle.set_color(.5, .5, .8)
-            self.viewer.add_geom(self.axle)
-            self.track = rendering.Line((0, carty), (screen_width, carty))
-            self.track.set_color(0, 0, 0)
-            self.viewer.add_geom(self.track)
-
-        if self.model.state is None:
+        Pyglet only supports multiple Displays on Linux.
+        """
+        if spec is None:
             return None
+        
+        if isinstance(spec, six.string_types):
+            return pyglet.canvas.Display(spec)
+        
+        raise ValueError("Invalid display specification: {}. \
+        (Must be a string like :0 or None.)".format(spec))
 
-        x = self.model.state
-        cartx = x[0]*scale+screen_width/2.0  # MIDDLE OF CART
-        self.carttrans.set_translation(cartx, carty)
-        self.poletrans.set_rotation(-x[2])
+    def update(self):
+        """ Update the game clock and dispatch game events to render the
+            cart.
+        """
+        pyglet.clock.tick()
+        self.switch_to()
+        self.dispatch_events()
+        self.dispatch_event('on_draw')
+        self.flip()
+    
+    def on_draw(self):
+        """ Draw the cartpole model in the window.
+        """
+        glClearColor(1, 1, 1, 1)
+        self.clear()
+        
+        draw_cartpole(self.model, self.width)
 
-        return self.viewer.render(return_rgb_array=mode == 'rgb_array')