+++ samples/stellar-cartpole-trpo/star.py
@@ -1,86 +1,54 @@
-"""
-This file contains machine teaching logic such as state, terminal, action,
-and reward functions.
+""" The state transform, action transform, terminal function and reward 
+    function. The functions in this file are called from cartpolesimulator.py.
 """
 
 import math
-import random
+
 from bonsai_ai.logger import Logger
 
 log = Logger()
 
-class Star():
-    iteration_count = 0
-    model_state = None
-    steps_beyond_done = None
 
-    def __init__(self):
-        # Angle at which to fail the episode
-        self.theta_threshold_radians = 12 * 2 * math.pi / 360
-        self.x_threshold = 2.4
+def state(model_state):
+    """ Convert the state as represented in CartPole to a format expected.
+        by the AI Engine.
+    """
+    return {
+        'position': model_state[0],
+        'velocity': model_state[1],
+        'angle': model_state[2],
+        'rotation': model_state[3],
+    }
 
-    def state(self, model_state):
-        """ This function converts the simulator state into the state
-        representation the brain uses to learn. """
-        self.model_state = model_state
-        brain_state = {
-            'position': model_state[0],
-            'velocity': model_state[1],
-            'angle': model_state[2],
-            'rotation': model_state[3],
-        }
-        print(brain_state)
-        return brain_state
 
-    def terminal(self, model_state):
-        """ Terminal conditions specify when to end an episode, typically due
-        to success, failure, or running out of time. In this case, we only
-        terminate when the pole falls down. """
-        print(model_state)
-        x, x_dot, theta, theta_dot = model_state
-        done = (x < -self.x_threshold or
-                x > self.x_threshold or
-                theta < -self.theta_threshold_radians or
-                theta > self.theta_threshold_radians)
-        done = bool(done)
-        return done
+def terminal(model_state):
+    """ Return true if the state should end the episode. This includes both
+        failure terminals (such as when the model isout-of-bounds) and success
+        terminals (when the model is in a successful state)
+    """
+    x, x_dot, theta, theta_dot = model_state
+    theta_threshold_radians = 12 * 2 * math.pi / 360
+    x_threshold = 2.4
+ 
+    # Terminal occurs when the cart's position is too far to the left or right
+    # or the cart's pole has tipped too far.
+    done = (x < -x_threshold or
+            x > x_threshold or
+            theta < -theta_threshold_radians or
+            theta > theta_threshold_radians)
+    return done
 
-    def action(self, brain_action):
-        """ This function converts the action representation the brain learns
-        to use into the action representation the simulator uses to act in the
-        simulated environment. """
-        self.iteration_count += 1
-        model_action = 0
-        if brain_action['command'] > 0:
-            model_action = 1
-        return model_action
+def action(brain_action):
+    """ Convert the action from the AI Engine to a format expected by the
+        CartPole model.
+    """
+    return 1 if brain_action['command'] > 0 else 0
 
-    def reward(self, done):
-        """ Give the brain feedback on how well it is doing in this episode.
-        In this case, this is simply 1 every time period that the pole is
-        balanced. The brain's job is to learn to maximize the reward it gets
-        during the episode, which corresponds to balancing as long as possible.
-        """
-        if not done:
-            reward = 1.0
-        elif self.steps_beyond_done is None:
-            # Pole just fell!
-            self.steps_beyond_done = 0
-            reward = 1.0
-        else:
-            if self.steps_beyond_done == 0:
-                log.info("You are calling 'step()' even though this \
-                         environment has already returned done = True. \
-                         You should always call 'reset()' once you receive \
-                         'done = True' -- any further steps are undefined \
-                         behavior.")
-            self.steps_beyond_done += 1
-            reward = 0.0
-        return reward
+def reward(model_state, done):
+    """ Return greater values to reward the AI for correct behavior.
+    """
+    # If the AI has not hit a terminal situation reward it with a score of 1.0
+    if not done:
+        return 1.0
+    return -0.01
 
-    def reset(self, model_state):
-        self.iteration_count = 0
-        brain_state = self.state(model_state)
-        self.steps_beyond_done = None
-
-        return brain_state